Parsing (ns)
This code isn't super well motivated - I was poking at clojure.tools.namespace.file and was
... dismayed to discover that it pretty much just returns the first form of the file which isn't
super useful when the goal for this project is to try and automatically understand a user's code
structure and then do something with it.

This namespace steals some implementation details from clojure.core and works fairly hard to
parse (ns) and the various use/require/import/gen-class forms in common usage to a declarative
data structure which should accurately predict the ns-map of the given namespace unless the user's
doing something with inline loads or other manual manipulation.

Realistically I'll just use this to extract the name of namespaces so that I can group files
together by the namespace they participate in and then try to figure out what files actually need to
be loaded to document a given thing. This could enable me to load some subset of code much more
efficiently using pods and understanding its dependency graph. Realistically I should just boot a
subprocess, load everything blindly and then work on inspecting it.